+++ src/frontend/components/ComponentOverviewColumns.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { ColumnDef, createColumnHelper } from '@tanstack/react-table'
-import { ComponentData } from './models'
+import { ComponentData, ComponentBoundingBox, ComponentPolylinePoints } from './models'
 import { rgbToHex } from '@/lib/utils'
 import ComponentSheet from './ComponentSheet'
 import ComponentPreviewImage from './ComponentPreviewImage'
@@ -17,17 +17,6 @@ export const ComponentOverviewColumns: ColumnDef<ComponentData>[] = [
           <ComponentSheet component_data={row.original} />
       )}
   },
-  // {
-  //   accessorKey: 'bbx',
-  //   header: () => <div className='align-top'>BBX</div>,
-  //   cell: ({ row }) => {
-  //     const component_bbx: string = row.original.bbx.xy.toString()
-  //   return (
-  //     <div className='text-left align-text-top'>
-  //       {component_bbx}
-  //     </div>
-  //   )}
-  // },
   {
     accessorKey: 'type',
     header: () => <div className='align-top'>Type</div>,
@@ -51,13 +40,53 @@ export const ComponentOverviewColumns: ColumnDef<ComponentData>[] = [
     )}
   },
   {
+    accessorKey: 'bbx',
+    header: 'X',
+    cell: ({ row }) => {
+      const component_bbx: ComponentBoundingBox = row.getValue('bbx')
+      const component_bbx_xy: ComponentPolylinePoints = component_bbx.xy
+      const component_bbx_xyz: ComponentPolylinePoints = component_bbx.xyz
+      let component_bbx_x: number = 0
+      if (component_bbx_xy) {
+        component_bbx_x = component_bbx_xy[1][0] - component_bbx_xy[0][0]
+      }
+      else {
+        component_bbx_x = component_bbx_xyz[1][0] - component_bbx_xyz[0][0]
+      }
+    return (
+      <div className='text-left align-text-top'>
+        {component_bbx_x.toFixed(2)}
+      </div>
+    )}
+  },
+  {
+    accessorKey: 'bbx',
+    header: 'Y',
+    cell: ({ row }) => {
+      const component_bbx: ComponentBoundingBox = row.getValue('bbx')
+      const component_bbx_xy: ComponentPolylinePoints = component_bbx.xy
+      const component_bbx_xyz: ComponentPolylinePoints = component_bbx.xyz
+      let component_bbx_y: number = 0
+      if (component_bbx_xy) {
+        component_bbx_y = component_bbx_xy[2][1] - component_bbx_xy[1][1]
+      }
+      else {
+        component_bbx_y = component_bbx_xyz[2][1] - component_bbx_xyz[1][1]
+      }
+    return (
+      <div className='text-left align-text-top'>
+        {component_bbx_y.toFixed(2)}
+      </div>
+    )}
+  },
+  {
     accessorKey: 'materialthickness',
     header: 'Z',
     cell: ({ row }) => {
       const component_z: number = Math.round(row.getValue('materialthickness'))
     return (
       <div className='text-left align-text-top'>
-        {component_z}
+        {component_z.toFixed(2)}
       </div>
     )}
   },
@@ -71,10 +100,21 @@ export const ComponentOverviewColumns: ColumnDef<ComponentData>[] = [
       const colB = Math.round(color[2])
       const hexcol = rgbToHex(colR, colG, colB)
       return (
-        <div className='flex items-center max-w-12'>
+        <div className='flex items-center'>
           <div className='avatar rounded-full min-h-4 min-w-4 max-w-4 max-h-4 items-center justify-left' style={{backgroundColor: hexcol}}></div>
           <div className='px-2 items-center justify-center text-center text'>{colR}/{colG}/{colB}</div>
         </div>
+    )}
+  },
+  {
+    accessorKey: 'created',
+    header: 'Created',
+    cell: ({ row }) => {
+      const creation_date: string = row.getValue('created')
+    return (
+      <div className='text-left align-text-top'>
+        {creation_date}
+      </div>
     )}
   },
 ]